Some option settings.
I followed the suggestion of the Github client and added its suggested
options. I think one of them was for suggested line endings. That was
probably a mistake as the client seemed to change the endings to cr/lf.
The option it selected was "text=auto" I have changed to "text=eol=lf"
which seems to have put the line endings back.

Moral of this story, be explicit about the line endings!